job-agent/pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "Standalone A2A Job Agent using Vertex AI Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langgraph>=0.3.34",
    "pydantic>=2.10.6",
    "langchain-google-vertexai>=2.0.21",
    "dotenv>=0.9.9",
    "jwcrypto>=1.5.6",
    "pyjwt>=2.10.1",
    "sse-starlette>=2.3.3",
    "uvicorn>=0.34.2",
    "a2a-sdk[http-server]>=0.2.16",
    "starlette>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


